pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "research-pipeline"
version = "0.1.0"
description = "Multi-stage AI pipeline that discovers research papers, extracts algorithms, and generates working prototypes"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.42.0",
    "arxiv>=2.1.0",
    "pymupdf>=1.25.0",
    "sympy>=1.13.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "hypothesis>=6.100",
    "ruff>=0.8.0",
]

[project.scripts]
research-pipeline = "research_pipeline.cli:cli"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
